PLF-7169 : Upgrade eXo Platform to ISPN last version (8)
