--- exo.kernel.component.ext.cache.impl.infinispan.v5/src/test/resources/test.policy
@@ -1,26 +1,0 @@
-grant codeBase "@MAVEN_REPO@-"{
-   permission java.security.AllPermission;
-};
-
-grant codeBase "@MAIN_CLASSES@-"{
-   permission java.security.AllPermission;
-};
-
-grant codeBase "@TEST_CLASSES@-"{
-};
-
-grant codeBase "@MAIN_CLASSES@../../../exo.kernel.commons.test/-"{
-   permission java.security.AllPermission;
-};
-
-grant codeBase "@MAIN_CLASSES@../../../exo.kernel.commons/-"{
-   permission java.security.AllPermission;
-};
-
-grant codeBase "@MAIN_CLASSES@../../../exo.kernel.container/-"{
-   permission java.security.AllPermission;
-};
-
-grant codeBase "@MAIN_CLASSES@../../../exo.kernel.component.cache/-"{
-   permission java.security.AllPermission;
-};
