+++ exo.kernel.component.ext.cache.impl.infinispan.v8/src/test/java/org/exoplatform/services/cache/impl/infinispan/TestExoCacheConfig.java
