+++ exo.kernel.component.ext.cache.impl.infinispan.v8/src/test/resources/conf/portal/test-configuration.xml
