+++ exo.kernel.component.ext.cache.impl.infinispan.v8/src/test/resources/tsm-excludes.properties
