+++ pom.xml
@@ -58,9 +58,9 @@
       <javax.transaction.version>1.1</javax.transaction.version>
       <connector-api.version>1.5</connector-api.version>
       <javax.servlet.version>3.0.1</javax.servlet.version>
-      <jgroups.version>3.2.7.Final</jgroups.version>
-      <jboss-logging.version>3.1.3.GA</jboss-logging.version>
-      <org.infinispan.version>5.2.9.Final</org.infinispan.version>
+      <jgroups.version>3.6.13.Final</jgroups.version>
+      <jboss-logging.version>3.3.0.Final</jboss-logging.version>
+      <infinispan.version>8.2.6.Final</infinispan.version>
       <org.ow2.jotm.version>2.1.9</org.ow2.jotm.version>
       <ow2-jta-1.1-spec.version>1.0.13</ow2-jta-1.1-spec.version>
       <transactions-jta.version>3.8.0</transactions-jta.version>
@@ -95,7 +95,7 @@
       <module>exo.kernel.commons.test</module>
       <module>exo.kernel.component.common</module>
       <module>exo.kernel.component.cache</module>
-      <module>exo.kernel.component.ext.cache.impl.infinispan.v5</module>
+      <module>exo.kernel.component.ext.cache.impl.infinispan.v8</module>
       <module>exo.kernel.component.ext.cache.impl.memcached.v1</module>
       <module>exo.kernel.component.command</module>
    </modules>
@@ -218,8 +218,13 @@
          <dependency>
             <groupId>org.infinispan</groupId>
             <artifactId>infinispan-core</artifactId>
-            <version>${org.infinispan.version}</version>
+            <version>${infinispan.version}</version>
          </dependency>
+          <dependency>
+              <groupId>org.infinispan</groupId>
+              <artifactId>infinispan-commons</artifactId>
+              <version>${infinispan.version}</version>
+          </dependency>
          <dependency>
             <groupId>org.jibx</groupId>
             <artifactId>jibx-bind</artifactId>
